feat: allow optional namespace configuration (#10)
This PR allows the ability to specify an optional namespace for all groups and permissions. This is useful so that they are easily distinguishable between different applications.

e.g. "app1:users.edit" and "app2:users.edit"

The namespace is set as part of the config in the initializer in the Rails application.

```ruby
Checken::Schema.instance.configure do |config|
  config.namespace = 'app1'
end
```

> [!NOTE]
> Additional details of namespace configuration are in the README.md